Implement KQStream#read, write unit tests
`KQStream#read(data: string)` processes messages from the `data` string. `data` must be the contents of a CSV with the following format:

- First value: timestamp in milliseconds
- Second value: message from the cabinet, not wrapped in quotes

The `#read` method simulates messages from a real Killer Queen cabinet. The first message in the `data` string is processed immediately, and processing of subsequent messages is delayed according to each message's timestamp.

Example of a valid `data` string:

```
1518063130451,![k[playerKill],v[730,860,9,8]]!
1518063130461,![k[playerKill],v[770,860,9,10]]!
1518063130471,![k[playerKill],v[810,860,9,4]]!
1518063130481,![k[playerKill],v[1071.977,20,3,10]]!
```

README has been updated, and unit tests have been written for the new method, as well as reorganized.