Implement OBSData class with example, fix KQStream bug
`OBSData` accepts an instance of the `KQStream` class and uses it to write files containing kill and death counts for every player. These files can be used by OBS to display the stats in a stream.

Each character has two files:

1. Kill count (single integer)
2. Death count (single integer)

All the values are reset to 0 when the `playernames` event is received from the cab. (This event indicates that a new game has started.)

Fixed a bug in `KQStream` where it would attempt to respond to an `alive` message from the cab when reading from a file. There would be no connection to a cab established, and sending a message would throw an error. (When a connection to a cab does exist, it will still send a message to the cab when the `alive` message is read from a file. This probably isn't harmful, but isn't the best behavior either...)