Track warrior kills
- `GameStats` class infers state from actions in the game, and uses this to track warrior kills
- Full killboard displays warrior kills

If a character gets a kill, you know they are a warrior. If that charater then gets killed, you know it was a warrior kill.

The approach is not 100% accurate. If a character becomes a warrior, gets no kills, then gets killed, we cannot know that it was a warrior kill.