Restructure some unit tests
**`KQCab`**

A number of these unit tests would resolve a promise when a certain number of calls were made. This works, but would also pass if more than the indicated number of calls were made. Also, structurally this approach isn't very sound.

To resolve this, we store an array of promises for these calls, and `Promise.all(promises)` at the end of the unit test, followed by assertions if relevant.

When using `Promise.all()`, we also include an assertion for the number of promises. If we call `Promise.all(promises)` but our `promises` array is empty, the test will still pass, so this serves as a sanity check.

Some unnecessary code was removed from some unit tests as well.

**`KQStream`**

Calls to `createListenerPromise()` now pass in the length of the expected events array. This allows us to add and remove events from these arrays without having to manually update the expected number of calls to the event listener.