Refactor KQStream unit tests
Removed `createListenerPromise()`. Now each message received by
`KQStream` is individually awaited, instead of awaiting a certain number
of messages to be received, which could be misleading.

For `KQStream#read` unit tests, `sleep-promise` is used. This is not
ideal and should be changed to use a different approach in the future.