Web app: single player stats, refactor
A new `Killboard` type was created that displays stats for a single player. Useful for players who want to see their stats on their phone as they play.

The homepage of kqstats now lets you pick a character, which brings you to the single player stats view.

Web app UI has been universally changed to white on black.

A number of things were refactored:

- `Killboard`: single class per file
- All sprites can now be imported at once, no need to `require()` the image files yourself
- Moved some files around

Added Bootstrap as a project dependency to speed up frontend development.

Updated some of `public/manifest.json` which were previously default. (`favicon.ico` is still default, and should be changed to a KQ spirte or something.)