Fix websocket connection
Previously the web app would connect to `http://localhost:8000` for websocket
messages from the server. This is a bad assumption and only works when
running the web app and server locally.

Now, it connects to `http://${window.location.hostname}:8000`. Regardless if you
visit the webpage via an IP address, `localhost`, or even
`kqstats.local` (or some other hostname), the websocket connection will work.