Autoconnect to cab
Running `npm run start` with no arguments will now look for IP addresses on the local network with port 12749 open. The first IP it finds with this port open is the one it will connect to.

This still doesn't handle disconnects from the cab, so it is not fully persistent.