Implement Game state machine, GameManager class
Introduces multiple new class:

- `Game` and the classes that it holds:
  - `Player`
  - `Gate`
  - `Snail`
  - `Hive`
- `GameManager`
- Numerous game `Error` classes

`Game` is a class that holds game state and acts as a state machine. To
change game state, public instance methods are called. The methods
perform a number of state checks, often through the classes it holds --
`Player`, `Gate`, `Snail`, and `Hive`. If these state checks pass, game
state is updated.

Example: when you call `Game#depositBerry`, a number of state checks are
performed:

- Do hives exist?
- Is the respective hive full?
- Is the depositing player a drone?
- Is the depositing player not on the snail?
- Does the depositing player have a berry?

When any of these state checks fails, a specific error type is thrown
(e.g. `DroneCanNotDepositBerryWhileOnSnailError`) to indicate the game
logic constraint that was not satisfied.

Some caveats about the `Game` class:

- `Game#setMap` will also set the gates, snails, and hives of the `Game`
  object
- Certain methods will infer and set the game map or other game elements
  automatically:
  - `Game#tagGate` and `Game#enterGate` use the position of the gate to
    set the game map
  - `Game#depositBerry` and `Game#kickInBerry` will create hives
  - `Game#getOnSnail` use the starting position of the snail (i.e. the
    first call to this method) to set the game map

`GameManager` processes messages from a `KQStream` object and calls the
respective methods on a `Game` object. It also emits a `Game` object when
a new game begins, which can then be consumed by other code.

The `Game` class is designed to be fully independent from `KQStream`,
hence the existence of the `GameManager` class to tie the two together.